refactor: 퀸, 룩, 비숍을 묶는 추상화 클래스 추가
