+++ src/main/java/chess/model/piece/Queen.java
@@ -9,14 +9,11 @@ import static chess.model.position.Direction.UP;
 import static chess.model.position.Direction.UP_LEFT;
 import static chess.model.position.Direction.UP_RIGHT;
 
-import chess.model.board.ChessBoard;
 import chess.model.board.Point;
-import chess.model.position.ChessPosition;
 import chess.model.position.Direction;
-import java.util.HashSet;
 import java.util.Set;
 
-public class Queen extends Piece {
+public class Queen extends JumpingPiece {
 
     private static final int QUEEN_POINT = 9;
 
@@ -25,39 +22,8 @@ public class Queen extends Piece {
     }
 
     @Override
-    public boolean isKing() {
-        return false;
-    }
-
-    @Override
-    public boolean isPawn() {
-        return false;
-    }
-
-    @Override
     public Point getPoint() {
         return Point.from(QUEEN_POINT);
-    }
-
-    @Override
-    protected Set<ChessPosition> calculatePaths(final ChessPosition source, final ChessPosition target,
-                                                final ChessBoard chessBoard) {
-        final Set<ChessPosition> paths = new HashSet<>();
-        Set<Direction> directions = availableDirections();
-        for (final Direction direction : directions) {
-            ChessPosition current = source;
-            while (current.canMove(direction)) {
-                current = current.move(direction);
-                if (chessBoard.isSameSide(current, side)) {
-                    break;
-                }
-                paths.add(current);
-                if (chessBoard.isEnemy(current, side)) {
-                    break;
-                }
-            }
-        }
-        return paths;
     }
 
     @Override
